requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
together==0.2.8
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0
prometheus-client==0.19.0
sentry-sdk==1.35.0
pytest==7.4.3
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1 